refactor: update source map folder name
