+++ packages/vite/src/node/ssr/runtime/sourcemap/index.ts
